Initial working version of containerised install
Removes all the bash nonsense from the source tree and uses Ansible to create a declaritive, idempotent install and uninstall.

Uses an ansible role to house tasks, variables and handlers that are common across playbooks (a lot more work needed on efficiency here).

This approach does away with the .env file and instead dynamically writes the service files to contain the required variable values directly. User specific config is now managed in a yaml file that is mounted into the container at runtime and used as an ansible inventory.

The installer (or uninstaller) can be run multiple times over the same targets and should be both safe and idempotent in all cases.