ci: autobump versions of containers
removes the need to manage (and remember to change) VERSION files for container images.

If changes are detected and the commit message is correct the version will be bumped to the
correct level automatically.

Use a commit message prefix of;

- `fix(scope):` to bump a patch version
- `feat(scope):` to bump a minor version
- `feat(scope)!:` to bump a major version

Biggest change wins if there are many and the VERSION file can also be manually adjusted if needed.